+++ packages/frontend/src/components/Notes/SearchResultList.vue
@@ -1,46 +1,37 @@
 <template>
   <div>
-    <div v-for="item in props.resultGroup" :key="item.note.id">
+    <div v-for="result in props.resultGroup" :key="result.item.id">
       <v-list-item
-        :to="noteUrl(item.note)"
+        :to="noteUrl(result.item)"
         :ripple="false"
         class="note-list-item"
       >
         <template #prepend>
           <div class="note-list-children-icon">
-            <v-icon v-if="item.children.length > 0" icon="mdi-menu-down" />
+            <v-icon v-if="result.children.length > 0" icon="mdi-menu-down" />
           </div>
           <div class="note-icon">
-            <v-icon v-if="item.note.checked === true" size="small" class="text-disabled" icon="mdi-checkbox-marked" />
-            <v-icon v-else-if="item.note.checked === false" size="small" class="text-disabled" icon="mdi-checkbox-blank-outline" />
-            <s-emoji v-else-if="item.note.icon_emoji" :value="item.note.icon_emoji" size="small" class="emoji-icon" />
-            <v-icon v-else-if="item.children.length > 0" size="small" class="text-disabled" icon="mdi-folder-outline" />
+            <v-icon v-if="result.item.checked === true" size="small" class="text-disabled" icon="mdi-checkbox-marked" />
+            <v-icon v-else-if="result.item.checked === false" size="small" class="text-disabled" icon="mdi-checkbox-blank-outline" />
+            <s-emoji v-else-if="result.item.icon_emoji" :value="result.item.icon_emoji" size="small" class="emoji-icon" />
+            <v-icon v-else-if="result.children.length > 0" size="small" class="text-disabled" icon="mdi-folder-outline" />
             <v-icon v-else size="small" class="text-disabled" icon="mdi-note-text-outline" />
           </div>
         </template>
         <template #default>
-          <v-list-item-title class="text-body-2">{{ item.note.title }}</v-list-item-title>
+          <v-list-item-title class="text-body-2">{{ result.item.title }}</v-list-item-title>
         </template>
       </v-list-item>
 
-      <v-list density="compact" class="match-list">
-        <v-list-item
-          v-for="match in item.matches" :key="match.field + match.from"
-          :to="noteUrl(item.note, match)"
-          @click="event => navigateToMatch(event, item.note, match)"
-          :active="false"
-        >
-          <v-list-item-title>
-            <span>{{ match.previewText.slice(0, match.previewFrom) }}</span>
-            <span class="bg-search-match">{{ match.previewText.slice(match.previewFrom, match.previewTo) }}</span>
-            <span>{{ match.previewText.slice(match.previewTo) }}</span>
-          </v-list-item-title>
-        </v-list-item>
-      </v-list>
+      <search-match-list
+        :result="result"
+        :to-prefix="noteUrl(result.item)"
+        class="match-list"
+      />
 
       <v-list density="compact" class="pt-0 pb-0">
         <notes-search-result-list
-          :result-group="item.children || []"
+          :result-group="result.children || []"
           :to-prefix="props.toPrefix"
           :level="(props.level || 0) + 1"
           class="child-list"
@@ -57,23 +48,9 @@ const props = defineProps<{
   level?: number;
 }>();
 
-const route = useRoute();
-
-function noteUrl(note: NoteBase, match?: SearchResultMatch) {
-  return `${props.toPrefix}${note.id}/` + (match ? `#${match.field}:offset=${match.from}` : '');
+function noteUrl(note: NoteBase,) {
+  return `${props.toPrefix}${note.id}/`;
 }
-
-function navigateToMatch(event: Event, note: NoteBase, match: SearchResultMatch) {
-  const url = new URL(noteUrl(note, match), window.location.href);
-  if (!url) {
-    return;
-  }
-  if (route.path === url.pathname) {
-    focusElement(url.hash, { scroll: { behavior: 'smooth', block: 'center' } });
-    event.preventDefault();
-  }
-}
-
 </script>
 
 <style lang="scss" scoped>
@@ -107,17 +84,5 @@ function navigateToMatch(event: Event, note: NoteBase, match: SearchResultMatch)
 
 .match-list {
   padding-left: 2rem;
-  padding-top: 0;
-  padding-bottom: 0;
-
-  .v-list-item {
-    padding-top: 0;
-    padding-bottom: 0;
-    min-height: 24px;
-
-    .v-list-item-title {
-      font-size: small;
-    }
-  }
 }
 </style>